refacto nuxt config + add translation + correct carousel + modif navbar
