+++ configs/server.js
@@ -1,0 +1,8 @@
+/* eslint-disable */
+
+require('dotenv').config()
+
+module.exports = {
+  port: process.env.PORT || 3000, // par défaut: 3000
+  host: '0.0.0.0' // par défaut: localhost
+}
